+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Changelog
 
   * output drivers must now provide `OutputDataWriter` and `OutputDataReader` classes instead of a single `OutputData` class
   * `OutputDataWriter.close()` method must accept `exc_type=None, exc_value=None, exc_traceback=None` keywords
+  * `mapchete pyramid` CLI was removed and is replaced by the more versatile `mapchete convert`
 
 * fixed `area_at_zoom()` when using input groups (#181)
 * fixed single GeoTIFF output bounds should use process area (#182)